When "native" slurm was first added, we couldn't get exclusive access to all
the network resources, so we had to limit how much memory we registered with
the NIC. This is no longer an issue, so we don't have to limit the number of
communication domains under slurm.

Note that we're still limiting the default heap size under slurm to 16GB
instead of 2/3 available memory. This could be changed too, but the smaller
size seems to be working for us, and it improves startup time. This is
something to revisit at another time.

Increasing the number of communication domains results in some pretty dramatic
performance improvements for some latency bound benchmarks. I'm seeing a 25%
speedup for SSCA, 50% speedup for Lulesh, and a 75% improvement for RA:

| Benchmark | Old Perf     | New Perf     |
| --------- | ------------ | ------------ |
| SSCA#2    | 111 sec      | 85 sec       |
| Lulesh    | 4.7 sec      | 3.3 sec      |
| RA-atomic | .075 GUPS    | .12 GUPS     |
| RA-rmo    | .040 GUPS    | .075 GUPS    |