@gbtitus 2017-03-14
This looks fine and the performance rebound is nice.  👍 

Either the task-qthreads change should be in a different PR or the title and message on this one should be adjusted, since that's not just a "while I was here" change.