@ronawho 2017-03-14
This is currently based on #5646 (which adds the qt spincont stuff), once that's merged I'll rebase.