@gbtitus 2017-03-15
Looks good -- okay to merge.